Improved the seeding and usage of rand
Seeding rand each time before use is slow, and it leads
to lower quality of randomness. Improved the seeding
by moving it out to `func init()`

Added the test `TestRandomStrElement` to demonstrate
how seeding with `time.Now().Unix()` can lead to generation
of the same value 100 times in a row.

There is an open issue on `func handleRolloutEpoch`.
It seems, the result `controlPlaneManager.internal.rolloutEpoch`
will be a random int64, negative in about 1/2 of cases

Signed-off-by: aviaconstructor <gh@aviaconstructor.ru>