Update method for enabling/disabling modules in documentation
Deleted the 'module-enable-snippet.liquid' file and updated the 'module-enable.liquid' file in the documentation to provide a more concise and straightforward explanation on how to enable or disable modules. This includes examples using both the ModuleConfig resource and the 'deckhouse-controller' command. The updated explanations are applicable in both English and Russian versions of the documentation.

Signed-off-by: Artem Kladov <artem.kladov@flant.com>