[Prometheus][memcached] Adding drop for memcached metrics (#8342)
* Adding drop for memcached metrics with name
"memcached_slab_*"

With increasing RAM consumption, the number of
"memcached_slab_*" metrics increases, for this
reason, "sample_limit" is exceeded. Metrics are
not used in monitoring.

* Adding drop for memcached metrics with name "memcached_slab_*"

With increasing RAM consumption, the number of "memcached_slab_*" metrics increases, for this reason, "sample_limit" is exceeded. Metrics are not used in monitoring.